Added some fixes
Fixed an issue where the Timer wasn't updating if the user updates their interest's length.
If the user starts the timer, pauses it, and updates anything other than the activity length, the time remaining will remain the same. Otherwise, the time remaining will reset to the new length.

Fixed an issue where if the number of notifications changes from high to low, that the database would still have values in those spots that are unused.

Added commenting.